+++ queries/vue/rainbow-tags.scm
@@ -1,0 +1,15 @@
+(element
+  (start_tag) @opening
+  (end_tag)   @closing) @container
+
+(template_element
+  (start_tag) @opening
+  (end_tag)   @closing) @container
+
+(script_element
+  (start_tag) @opening
+  (end_tag)   @closing) @container
+
+(style_element
+  (start_tag) @opening
+  (end_tag)   @closing) @container