+++ CHANGELOG.rst
@@ -18,6 +18,7 @@ Added
 - Compatibility with Tree-sitter API in Neovim 0.9
 - Support for R (`r`)
 - Go queries for nodes `interface_type`, `map_type` and `var_declaration`
+- Query `rainbow-tags` for Vue.js
 
 Fixed
 =====